+++ src/commands/index.named.ts
@@ -1,4 +1,6 @@
 export * from './convert';
 export * from './convertSelection';
 export * from './pasteClipboard';
-export * from './reverseConvertSelection';
+export * from './deconvert';
+export * from './deconvertSelection';
+export * from './deconvertSelectionDelimiter';